Load OBJ model

+++ test.cpp
@@ -402,7 +402,7 @@ protected:
     return texture_;
   }
   const std::vector<vka::Vertex> vertices() { return vertices_; }
-  const std::vector<uint16_t> indices() { return indices_; }
+  const std::vector<uint32_t> indices() { return indices_; }
   const vka::UniformBufferObject uniform_buffer_object() {
     return uniform_buffer_object_;
   }
@@ -467,7 +467,7 @@ private:
       {{0.5f, -0.5f, -0.5f}, {0.0f, 1.0f, 0.0f}, {1.0f, 0.0f}},
       {{0.5f, 0.5f, -0.5f}, {0.0f, 0.0f, 1.0f}, {1.0f, 1.0f}},
       {{-0.5f, 0.5f, -0.5f}, {1.0f, 1.0f, 1.0f}, {0.0f, 1.0f}}};
-  const std::vector<uint16_t> indices_ = {0, 1, 2, 2, 3, 0, 4, 5, 6, 6, 7, 4};
+  const std::vector<uint32_t> indices_ = {0, 1, 2, 2, 3, 0, 4, 5, 6, 6, 7, 4};
   vka::Texture texture_;
   const vka::UniformBufferObject uniform_buffer_object_ = {};
   WindowManager window_manager_ = WindowManager();
